Make DatabaseThingy API consistent and more intuitive
From now on, DatabaseThingy single-underscore-prefixed attributes
(i.e. _database, _table, _database_name and _table_name) will always have
precedence in the database/table resolution.

By doing so, we are forced to do the database/table resolution again every
time, since we cannot (or don't want to) use those attributes as cache anymore;
but it's a small price to pay for a greater purpose.

This commit fixes #6.